// src/generated/models/Cr38e_purchaserequestapprovalsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr38e_purchaserequestapprovalscr38e_approvalstatus = {
  357090000: 'PendingApproval',
  357090001: 'Approved',
  357090002: 'Rejected',
  357090003: 'FuturePending',
  357090004: 'Skipped'
} as const;
export type Cr38e_purchaserequestapprovalscr38e_approvalstatus = keyof typeof Cr38e_purchaserequestapprovalscr38e_approvalstatus;
export const Cr38e_purchaserequestapprovalscr38e_status = {
  0: 'Rejected',
  1: 'Pending',
  2: 'FuturePending',
  3: 'Skipped',
  4: 'Approved'
} as const;
export type Cr38e_purchaserequestapprovalscr38e_status = keyof typeof Cr38e_purchaserequestapprovalscr38e_status;
export const Cr38e_purchaserequestapprovalsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr38e_purchaserequestapprovalsstatecode = keyof typeof Cr38e_purchaserequestapprovalsstatecode;
export const Cr38e_purchaserequestapprovalsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr38e_purchaserequestapprovalsstatuscode = keyof typeof Cr38e_purchaserequestapprovalsstatuscode;

export interface Cr38e_purchaserequestapprovalsBase {
  cr38e_approvalcomment?: string;
  cr38e_approvaldate?: string;
  cr38e_approvalname: string;
  cr38e_approvalstatus: Cr38e_purchaserequestapprovalscr38e_approvalstatus;
  cr38e_approvedby?: string;
  cr38e_approver: string;
  cr38e_level: number;
  "cr38e_PurchaseRequest@odata.bind": string;
  cr38e_purchaserequestapprovalid: string;
  cr38e_status: Cr38e_purchaserequestapprovalscr38e_status;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr38e_purchaserequestapprovalsstatecode;
  statuscode?: Cr38e_purchaserequestapprovalsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr38e_purchaserequestapprovals extends Cr38e_purchaserequestapprovalsBase {
  cr38e_approvalstatusname?: string;
  cr38e_purchaserequestname?: string;
  cr38e_statusname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr38e_purchaserequest?: object;
  _cr38e_purchaserequest_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}